requirements.txt
streamlit
google-generativeai
python-dotenv
pydantic_settings